Fix logical errors in SubsystemManager
The SubsystemManager, when extending IterSubsystem, could add itself to
the subsystem list which would allow it to call its own loop function
when a loop function would be called (which would call the loop
function... forever). By having the SubsystemManager be its own static
class that does not extend IterSubsystem fixes this problem.