Following similar to trajectory
Robot responds similarly to as if it was following the path, ei. it
turns on the circle when commanded to turn on the circle vs going
forwards. Changes include: fixing encoder clicks in tool and
recharacterization and changes to start and end velocity (in
TBase.java). Needed changes include: fixing the output distance and
velocity, making changes to TBase.java to make more things be built into
it.